Changed utils  method
