+++ src/test/java/appsCommon/PlaywrightUtils/Playwright_Common_Utils.java
@@ -1,7 +1,12 @@
 package appsCommon.PlaywrightUtils;
 
 import com.microsoft.playwright.Locator;
+import com.microsoft.playwright.Page;
+import com.microsoft.playwright.options.AriaRole;
+import com.nci.automation.utils.MiscUtils;
 import com.nci.automation.web.PlaywrightUtils;
+
+import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
 
 public class Playwright_Common_Utils {
 
@@ -13,5 +18,18 @@ public class Playwright_Common_Utils {
     public static void scrollIntoView(String locator){
         Locator element = PlaywrightUtils.page.locator(locator);
         element.evaluate("element => element.scrollIntoView()");
+    }
+
+    /**
+     * asserting the new page link and footer.
+     *
+     * @param name
+     * @param NewPage
+     */
+    public static void assertNewPage(String Name, String NewPage){
+        Page newPage1 = PlaywrightUtils.context.waitForPage(() -> {
+            PlaywrightUtils.page.getByRole(AriaRole.LINK, new Page.GetByRoleOptions().setName(Name)).click();});
+        MiscUtils.sleep(2000);
+        assertThat(newPage1).hasURL(NewPage);newPage1.close();
     }
 }
